Improve, expand PlotFindSampleSizes. Add alt methods to FindSampleSize Remove SampleFn interface. Remove sampleMean, sampleCount from GenSampleFn.
